- renamed RealmLiveDataList to LiveRealmResults and made it sortable - reworked add / delete exercise in repo - split onSelect from spinner binding adapter - fixed AFragment to use always UIThread for Toasts - added Toolbar to ExerciseFragment with sort and search function - added FilterRealmResultsObserver - fixed minor bugs in LiveDataList and added removeRange function - added search and sort functionality to ExerciseViewModel

+++ app/src/main/java/de/itdude/gymdude/viewmodel/ExerciseViewModel.kt
@@ -1,32 +1,44 @@
 package de.itdude.gymdude.viewmodel
 
+import androidx.appcompat.widget.SearchView
 import de.itdude.gymdude.BR
 import de.itdude.gymdude.R
+import de.itdude.gymdude.repo.db.LiveRealmResults
 import de.itdude.gymdude.repo.db.model.BodyPart
 import de.itdude.gymdude.repo.db.model.Exercise
+import de.itdude.gymdude.util.FilterRealmResultsObserver
 import de.itdude.gymdude.util.LiveDataList
 import de.itdude.gymdude.util.daysAgo
+import io.realm.Sort
 import javax.inject.Inject
 
-class ExerciseViewModel @Inject constructor() : AViewModel() {
+class ExerciseViewModel @Inject constructor() : AViewModel(), SearchView.OnQueryTextListener {
 
     // provides RecyclerView the fields where to put created bindings (via BindingAdapter)
     val viewModelBinding: Int = BR.vm
     val itemBinding: Int = BR.exercise
 
-    lateinit var exercises: LiveDataList<Exercise>
+    val exercises = LiveDataList<Exercise>()
+    private val filterObserver = FilterRealmResultsObserver(exercises) {it.name.contains(query)}
+    private var query = ""
+    private lateinit var results: LiveRealmResults<Exercise>
 
     override fun onCreate() {
-        exercises = repo.getExercises()
+        results = repo.getExercises()
+        results.sort("name")
+        results.observeForever(filterObserver)
     }
 
     fun addExercise() {
-        val newExercise = Exercise("Bankdrücken ${exercises.size}", BodyPart("Brust"))
+        val newExercise = Exercise("Bankdrücken ${results.size}", BodyPart("Brust"))
         repo.addExercise(newExercise) { error -> showToast(error) }
     }
 
-    fun deleteExercise(exercise: Exercise) =
-        repo.deleteExercise(exercise) { error -> showToast(error) }
+    fun deleteExercise(exercise: Exercise) {
+        if (exercise.isValid) {
+            repo.deleteExercise(exercise) { error -> showToast(error) }
+        }
+    }
 
     fun lastTimeDoneText(exercise: Exercise): String =
         when (val days = exercise.getLastTimeDone()?.daysAgo()) {
@@ -35,4 +47,19 @@ class ExerciseViewModel @Inject constructor() : AViewModel() {
             1 -> resources.getString(R.string.yesterday_text)
             else -> resources.getString(R.string.days_ago_text, days.toString())
         }
+
+    override fun onQueryTextSubmit(query: String?) = true
+
+    override fun onQueryTextChange(newText: String?): Boolean {
+        query = newText ?: ""
+        results.notifyObserver()
+        return true
+    }
+
+    override fun onCleared() = super.onCleared().also {
+        results.removeObserver(filterObserver)
+    }
+
+    fun sort(fieldName: String, order: Sort) = results.sort(fieldName, order)
+
 }
